enhancement: adds CacheHandler
